Renamed actionlog to tasklog to better capture the intent that a task can be checked off by default if it was not necessary for the particular deposit-bag
