+++ src/main/java/nl/knaw/dans/dvingest/core/bagprocessor/DatasetVersionCreator.java
@@ -22,8 +22,8 @@ import nl.knaw.dans.dvingest.core.dansbag.exception.RejectedDepositException;
 import nl.knaw.dans.dvingest.core.service.DataverseService;
 import nl.knaw.dans.dvingest.core.yaml.Expect;
 import nl.knaw.dans.dvingest.core.yaml.Init;
-import nl.knaw.dans.dvingest.core.yaml.actionlog.CompletableItem;
-import nl.knaw.dans.dvingest.core.yaml.actionlog.InitLog;
+import nl.knaw.dans.dvingest.core.yaml.tasklog.CompletableItem;
+import nl.knaw.dans.dvingest.core.yaml.tasklog.InitLog;
 import nl.knaw.dans.lib.dataverse.DataverseException;
 import nl.knaw.dans.lib.dataverse.model.RoleAssignment;
 import nl.knaw.dans.lib.dataverse.model.RoleAssignmentReadOnly;
