+++ src/test/java/nl/knaw/dans/dvingest/core/bagprocessor/DatasetVersionCreatorTest.java
@@ -21,8 +21,8 @@ import nl.knaw.dans.dvingest.core.service.DataverseService;
 import nl.knaw.dans.dvingest.core.service.YamlService;
 import nl.knaw.dans.dvingest.core.service.YamlServiceImpl;
 import nl.knaw.dans.dvingest.core.yaml.InitRoot;
-import nl.knaw.dans.dvingest.core.yaml.actionlog.CompletableItem;
-import nl.knaw.dans.dvingest.core.yaml.actionlog.InitLog;
+import nl.knaw.dans.dvingest.core.yaml.tasklog.CompletableItem;
+import nl.knaw.dans.dvingest.core.yaml.tasklog.InitLog;
 import nl.knaw.dans.lib.dataverse.model.RoleAssignmentReadOnly;
 import nl.knaw.dans.lib.dataverse.model.dataset.Dataset;
 import org.junit.jupiter.api.BeforeEach;
